NP-3460: Moved some api client logic to a class that can be reused across modules
